tgnpdcl-systems/tgnpdcl-document-service/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=20.1
boto3>=1.28  # For S3 support if needed later
